% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\docType{package}
\name{chromVARmotifs}
\alias{chromVARmotifs}
\alias{chromVARmotifs-package}
\title{chromVARmotifs: Collections of motifs for use with chromVAR}
\description{
Collections of motifs in a format for easy use with motifmatchr and chromVAR
packages.
}
