% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\docType{data}
\name{encode_pwms}
\alias{encode_pwms}
\title{encode_pwms}
\usage{
data(encode_pwms)
}
\value{
\code{\link[TFBSTools]{XMatrixList}} of length 1346
}
\description{
Collection of motifs from ENCODE project
}
\details{
Position frequency matrices were converted to PWMs by taking the log
of the frequencies (after adding a pseudocount of 0.008) divided by 0.25.
}
\examples{
data(encode_pwms)
}
\keyword{datasets}
