% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\docType{data}
\name{human_pwms_v2}
\alias{human_pwms_v2}
\title{human_pwms_v2}
\usage{
data(human_pwms_v1)
}
\value{
\code{\link[TFBSTools]{XMatrixList}} of length 870
}
\description{
Collection of human pwms filtered from version 1
}
\details{
These motifs were curated from the cisBP database. Position
frequency matrices were converted to PWMs by taking the log
of the frequencies (after adding a pseudocount of 0.008) divided by 0.25.
}
\examples{
data(human_pwms_v2)
}
\keyword{datasets}
