% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\docType{data}
\name{mouse_pwms_v2}
\alias{mouse_pwms_v2}
\title{mouse_pwms_v2}
\usage{
data(mouse_pwms_v2)
}
\value{
\code{\link[TFBSTools]{XMatrixList}} of length 884
}
\description{
Collection of mouse pwms filtered from version 1
}
\details{
These motifs were curated from the cisBP database. Position
frequency matrices were converted to PWMs by taking the log
of the frequencies (after adding a pseudocount of 0.008) divided by 0.25.
}
\examples{
data(mouse_pwms_v2)
}
\keyword{datasets}
